test: move fixture to conftest
